pyproject.toml
[tool.poetry]
name = "plushie-python"
version = "0.1.0"
description = "Code for the plushie project"
authors = ["Milosz Dobosz <ex19346@qmul.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ollama = "^0.2.0"
piper-tts = "^1.2.0"
sounddevice = "^0.4.6"
openai-whisper = "^20231117"
rich = "^13.7.1"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
